aoc-10: tubes
I WAY overthought this one to start. When I read the description the
first time I just kind of thought the intent was to prevent the solution
from being a basic fill, so I put it out of my mind. Turned out the
solution was as simple as filling in adjacent tiles, just with an extra
step.